Starting to normalize the project

+++ IntelliJ Project/src/main/java/com/globant/musicstore/dao/InvoiceDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.globant.musicstore.entity.Invoice;
 
-public interface InvoiceDAO extends JpaRepository<Invoice, Integer>{
+public interface InvoiceDAO extends JpaRepository<Invoice, Long>{
 
 }
